Split be::engine module
